Conflicts:
	3rdParty/init.py
	applications/sam/src/sam/diff/main.cpp

+++ 3rdParty/init.py
@@ -76,27 +76,34 @@ def getKnownExtensions( filename ):
 	return [f for f in knownExtensions.keys() if filename.endswith(f)]
 
 def uncompress(filename, ext, inNewDirectory, libname, folderExtracted):
-
 	if ext == 'tar.gz' :
 		tar = tarfile.open( download_dir + "/" + filename, 'r:*')
 		folder = './'
 		tar.extractall( folder )
+		if not folderExtracted :
+			folderExtracted = tar.getnames()[0]
 
 	if ext == 'tar.bz2' :
 		tar = tarfile.open( download_dir + "/" + filename, 'r:*')
 		folder = './'
 		tar.extractall( folder )
+		if not folderExtracted :
+			folderExtracted = tar.getnames()[0]
 
 	if ext == 'tgz' :
 		tar = tarfile.open( download_dir + "/" + filename, 'r:*')
 		folder = './'
 		tar.extractall( folder )
+		if not folderExtracted :
+			folderExtracted = tar.getnames()[0]
 
 	if ext == 'zip' :
 		zip = zipfile.ZipFile( download_dir + "/" + filename, 'r' )
 		folder = './'
 		if inNewDirectory == True : folder += filename[:-len(ext)-1]
 		zip.extractall( folder )
+		if not folderExtracted :
+			folderExtracted = zip.namelist()[0]
 
 	if ext == 'exe' :
 		fileToRun = download_dir + "/" + filename + ".exe"
@@ -151,30 +158,33 @@ def getAndUncompress( libraries ):
 			print ( 'uncompress error (%s)' % str(e) )
 
 allLibs = [
-		('zlib',        'http://prdownloads.sourceforge.net/libpng/zlib-1.2.3.tar.gz'                                               , 'zlib-1.2.3'                     , False), #if not windows else ('zlib'    , 'http://downloads.sourceforge.net/project/libpng/zlib/1.2.3/zlib123.zip'                      , '', True),
-		('libpng',      'http://prdownloads.sourceforge.net/libpng/libpng-1.2.41.tar.gz'                                            , 'libpng-1.2.41'                  , False), #if not windows else ('libpng'  , 'http://downloads.sourceforge.net/project/libpng/libpng12/older-releases/1.2.41/lpng1241.zip' , '', False),
-		('ilmbase',     'http://download.savannah.nongnu.org/releases/openexr/ilmbase-1.0.1.tar.gz'                                 , 'ilmbase-1.0.1'                  , False),
-		('openexr',     'http://download.savannah.nongnu.org/releases/openexr/openexr-1.6.1.tar.gz'                                 , 'openexr-1.6.1'                  , False),
-		('ctl',         'http://sourceforge.net/projects/ampasctl/files/ctl/ctl-1.4.1/ctl-1.4.1.tar.gz/download'                    , 'ctl-1.4.1'                      , False),
-		('boost',       'http://prdownloads.sourceforge.net/boost/boost_1_49_0.tar.bz2'                                             , 'boost_1_49_0'                   , False), #if not windows else ('boost'   , 'http://sourceforge.net/projects/boost/files/boost/1.47.0/boost_1_47_0.zip'                   , '', False),
-		('freetype',    'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                                         , 'freetype-2.4.3'                 , False), #if not windows else ('freetype', 'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                           , '', False),
-		('libraw',      'http://www.libraw.org/data/LibRaw-0.14.5.tar.gz'                                                           , 'LibRaw-0.14.5'                  , False), #if not windows else ('libraw'  , 'http://www.libraw.org/data/LibRaw-0.13.5.zip'                                                , '', False),
-		('libcaca',     'http://caca.zoy.org/files/libcaca/libcaca-0.99.beta17.tar.gz'                                              , 'libcaca-0.99.beta17'            , False),
-		('lcms2',       'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.tar.gz'                                   , 'lcms2-2.1'                      , False), #if not windows else ('lcms'    , 'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.zip'                        , '', False),
-		('bzlib',       'http://www.bzip.org/1.0.6/bzip2-1.0.6.tar.gz'                                                              , 'bzip2-1.0.6'                    , False),
-		('libtiff',     'ftp://ftp.remotesensing.org/pub/libtiff/tiff-3.9.5.tar.gz'                                                 , 'tiff-3.9.5'                     , False), #if not windows else ('libtiff' , 'ftp://ftp.remotesensing.org/pub/libtiff/tiff-3.9.5.zip'                                      , '', False),
-		('libtool',     'ftp://ftp.gnu.org/gnu/libtool/libtool-2.4.tar.gz'                                                          , 'libtool-2.4'                    , False),
-		('libxml',      'ftp://xmlsoft.org/libxml2/libxml2-2.7.8.tar.gz'                                                            , 'libxml2-2.7.8'                  , False), #if not windows else ('libxml'  , 'ftp://xmlsoft.org/libxml2/win32/libxml2-2.7.7.win32.zip'                                     , '', False),
-		('jasper',      'http://www.ece.uvic.ca/~mdadams/jasper/software/jasper-1.900.1.zip'                                        , 'jasper-1.900.1'                 , False),
-		('xz',          'http://tukaani.org/xz/xz-5.0.2.tar.gz'                                                                     , 'xz-5.0.2'                       , False),
-		('gvc',         'http://www.graphviz.org/pub/graphviz/stable/SOURCES/graphviz-2.26.3.tar.gz'                                , 'graphviz-2.26.3'                , False),
-		('jpeg',        'http://freefr.dl.sourceforge.net/project/libjpeg/libjpeg/6b/jpegsrc.v6b.tar.gz'                            , 'jpeg-6b'                        , False), #if not windows else ('jpeg'    , 'http://netcologne.dl.sourceforge.net/project/libjpeg/libjpeg/6b/jpegsr6.zip'                 , '', False),
-		('turbojpeg',   'http://garr.dl.sourceforge.net/project/libjpeg-turbo/1.2.0/libjpeg-turbo-1.2.0.tar.gz'                     , 'libjpeg-turbo-1.2.0'            , False),
-		('openjpeg',    'http://openjpeg.googlecode.com/files/openjpeg_v1_4_sources_r697.tgz'                                       , 'openjpeg_v1_4_sources_r697'     , False),
-		('ffmpeg',      'http://ffmpeg.org/releases/ffmpeg-0.8.2.tar.bz2'                                                           , 'ffmpeg-0.8.2'                   , False),
-		('openimageio', 'https://github.com/OpenImageIO/oiio/tarball/RB-1.0/OpenImageIO-oiio-Release-1.0.0-0-g5b37f1c.tar.gz'       , 'OpenImageIO-oiio-6441966'       , False),
-		('opencolorio', 'http://github.com/imageworks/OpenColorIO/tarball/v1.0.6/imageworks-OpenColorIO-v1.0.6-0-g0f89532.tar.gz'   , 'imageworks-OpenColorIO-a16d9ac' , False),
-		('imagemagick', 'http://www.imagemagick.org/download/ImageMagick.tar.gz'                                                    , 'ImageMagick-6.7.5-10'            , False)
+		('zlib',        'http://prdownloads.sourceforge.net/libpng/zlib-1.2.3.tar.gz'                                               , '', False), #if not windows else ('zlib'    , 'http://downloads.sourceforge.net/project/libpng/zlib/1.2.3/zlib123.zip'                      , '', True),
+		('libpng',      'http://prdownloads.sourceforge.net/libpng/libpng-1.2.41.tar.gz'                                            , '', False), #if not windows else ('libpng'  , 'http://downloads.sourceforge.net/project/libpng/libpng12/older-releases/1.2.41/lpng1241.zip' , '', False),
+		('ilmbase',     'http://download.savannah.nongnu.org/releases/openexr/ilmbase-1.0.1.tar.gz'                                 , '', False),
+		('openexr',     'http://download.savannah.nongnu.org/releases/openexr/openexr-1.6.1.tar.gz'                                 , '', False),
+		('ctl',         'http://sourceforge.net/projects/ampasctl/files/ctl/ctl-1.4.1/ctl-1.4.1.tar.gz/download'                    , '', False),
+		('boost',       'http://prdownloads.sourceforge.net/boost/boost_1_49_0.tar.bz2'                                             , '', False), #if not windows else ('boost'   , 'http://sourceforge.net/projects/boost/files/boost/1.47.0/boost_1_47_0.zip'                   , '', False),
+		('freetype',    'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                                         , '', False), #if not windows else ('freetype', 'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                           , '', False),
+		('libraw',      'http://www.libraw.org/data/LibRaw-0.14.5.tar.gz'                                                           , '', False), #if not windows else ('libraw'  , 'http://www.libraw.org/data/LibRaw-0.13.5.zip'                                                , '', False),
+		('libcaca',     'http://caca.zoy.org/files/libcaca/libcaca-0.99.beta17.tar.gz'                                              , '', False),
+		('lcms2',       'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.tar.gz'                                   , '', False), #if not windows else ('lcms'    , 'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.zip'                        , '', False),
+		('bzlib',       'http://www.bzip.org/1.0.6/bzip2-1.0.6.tar.gz'                                                              , 'bzip2-1.0.6', False),
+		('libtiff',     'ftp://ftp.remotesensing.org/pub/libtiff/tiff-3.9.5.tar.gz'                                                 , '', False), #if not windows else ('libtiff' , 'ftp://ftp.remotesensing.org/pub/libtiff/tiff-3.9.5.zip'                                      , '', False),
+		('libtool',     'ftp://ftp.gnu.org/gnu/libtool/libtool-2.4.tar.gz'                                                          , '', False),
+		('libxml',      'ftp://xmlsoft.org/libxml2/libxml2-2.7.8.tar.gz'                                                            , '', False), #if not windows else ('libxml'  , 'ftp://xmlsoft.org/libxml2/win32/libxml2-2.7.7.win32.zip'                                     , '', False),
+		('jasper',      'http://www.ece.uvic.ca/~mdadams/jasper/software/jasper-1.900.1.zip'                                        , '', False),
+		('xz',          'http://tukaani.org/xz/xz-5.0.2.tar.gz'                                                                     , '', False),
+		('gvc',         'http://www.graphviz.org/pub/graphviz/stable/SOURCES/graphviz-2.26.3.tar.gz'                                , '', False),
+		('jpeg',        'http://freefr.dl.sourceforge.net/project/libjpeg/libjpeg/6b/jpegsrc.v6b.tar.gz'                            , '', False), #if not windows else ('jpeg'    , 'http://netcologne.dl.sourceforge.net/project/libjpeg/libjpeg/6b/jpegsr6.zip'                 , '', False),
+		('turbojpeg',   'http://garr.dl.sourceforge.net/project/libjpeg-turbo/1.2.0/libjpeg-turbo-1.2.0.tar.gz'                     , '', False),
+		('openjpeg',    'http://openjpeg.googlecode.com/files/openjpeg_v1_4_sources_r697.tgz'                                       , '', False),
+		('ffmpeg',      'http://ffmpeg.org/releases/ffmpeg-0.8.2.tar.bz2'                                                           , '', False),
+		('openimageio', 'https://github.com/OpenImageIO/oiio/tarball/RB-1.0/OpenImageIO-oiio-Release-1.0.0-0-g5b37f1c.tar.gz'       , '', False),
+		('opencolorio', 'http://github.com/imageworks/OpenColorIO/tarball/v1.0.6/imageworks-OpenColorIO-v1.0.6-0-g0f89532.tar.gz'   , '' , False),
+		('imagemagick', 'http://www.imagemagick.org/download/ImageMagick.tar.gz'                                                    , '', False),
+		('webp',        'https://webp.googlecode.com/files/libwebp-0.1.3.tar.gz'                                                    , '', False),
+		('tinyxml',     'http://netcologne.dl.sourceforge.net/project/tinyxml/tinyxml/2.6.1/tinyxml_2_6_1.tar.gz'                   , '', False),
+		('yaml',        'http://yaml-cpp.googlecode.com/files/yaml-cpp-0.2.6.tar.gz'                                                , '', False),
 	]
 
 
@@ -198,21 +208,25 @@ def insertInFile( filename, linesIndexes, textToAppend ):
 	os.rename('tmp.txt', filename)
 
 def commentInFile( filename, atLine ):
-	file = open( filename , 'r' )
-	fileTmp = open( 'tmp.txt' , 'w' )
-	lines = file.readlines()
+	print ( '%s' % filename )
+	if(os.path.exists( filename )):
+		file = open( filename , 'r' )
+		fileTmp = open( 'tmp.txt' , 'w' )
+		lines = file.readlines()
 
-	nlist = list( lines )
+		nlist = list( lines )
 
-	commentedLine = nlist[atLine]
-	commentedLine = "//" + commentedLine
-	nlist[atLine]=commentedLine
-	fileTmp.writelines( nlist )
+		commentedLine = nlist[atLine]
+		commentedLine = "//" + commentedLine
+		nlist[atLine]=commentedLine
+		fileTmp.writelines( nlist )
 
-	file.close()
-	fileTmp.close()
-	os.remove(filename)
-	os.rename('tmp.txt', filename)
+		file.close()
+		fileTmp.close()
+		os.remove(filename)
+		os.rename('tmp.txt', filename)
+	else:
+		print( 'file is not present.' )
 
 
 def makeModificationIfNecessaryInFile( filename, headerToInsert, atLine ):
@@ -253,12 +267,7 @@ headerToInsert = "#include <cstring>"
 atLine = 55
 makeModificationIfNecessaryInFile(filename, headerToInsert, atLine );
 
-#OpenColorIO third parties extraction
-
-download_dir = "opencolorio/ext"
-
-uncompress("tinyxml_2_6_1.tar.gz", "tar.gz", False, "tinyxml", "tinyxml")
-uncompress("yaml-cpp-r482.tar.gz", "tar.gz", False, "yaml", "yaml-cpp-r482")
+#OpenColorIO third parties patch
 
 #patch yaml
 filename = "yaml/src/emitter.cpp"
